AnalysisManager: Generalize to multiple IRUnitT's
This should theoretically be able to handle multiple IRUnitT's in the
same analysis manager. But this is currently untested. To avoid needing
to churn all the new PM transformations / analyses, we keep multiple
separate analysis managers in this patch.
An upcoming patch will actually properly unify them (which will have a
lot of churn).

The basic approach here has been to change the IRUnitT template argument
to AnalysisManager{,Base} to IRUnitTOnClass and add templating to the
methods instead.

Simultaneously, we need to move up from a `void *` AnalysisID (sometimes
calls just PassID in the previous code) to a struct that also tracks the
IRUnitT (since the same analysis class can have an overloaded `run`
method for multiple different types (e.g. TargetLibraryInfo)).
For now, we just track an integer ID for each IRUnitT.
The exact integer is defined by a few functions like:
static inline int getIRUnitKindID(Function *) { return 0; }
static inline int getIRUnitKindID(Module *) { return 1; }
This is quick and dirty. I wouldn't be surprised if there is a cleaner
way (e.g. using the address of a global `char` as identification).

This passes check-llvm at least.