[PM] Port everything to the unified analysis manager.
This ports everything to the unified analysis manager.
The main "FIXME" on this is that currently the downward invalidation
only works for Function's and Module's (and it can't happen for
Modules).
Doing it for Loop and SCC would be a layering violation.
We can do this with an indirect interface, but that would require that
every user of the analysis manager register the appropriate handlers for
those two IRUnitT's; I'm open to better ideas.
Some users, like unittests/IR/PassManagerTest.cpp wouldn't be able to
register them, for example (since that test can't depend on Analysis).

Also, we add some dummy typedef's for the analysis managers to reduce
the source churn:
FunctionAnalysisManager, ModuleAnalysisManager, LoopAnalysisManager,
CGSCCAnalysisManager
They can trivially be renamed to AnalysisManager later.